refactor(examples): Update to use new response formats and packages
