--- examples/src/main/java/com/ibm/watson/developer_cloud/speech_to_text/v1/MicrophoneWithWebSocketsExample.java
@@ -1,82 +1,0 @@
-/*
- * Copyright 2017 IBM Corp. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package com.ibm.watson.developer_cloud.speech_to_text.v1;
-
-import javax.sound.sampled.AudioFormat;
-import javax.sound.sampled.AudioInputStream;
-import javax.sound.sampled.AudioSystem;
-import javax.sound.sampled.DataLine;
-import javax.sound.sampled.TargetDataLine;
-
-import com.ibm.watson.developer_cloud.http.HttpMediaType;
-import com.ibm.watson.developer_cloud.speech_to_text.v1.model.RecognizeOptions;
-import com.ibm.watson.developer_cloud.speech_to_text.v1.model.SpeechRecognitionResults;
-import com.ibm.watson.developer_cloud.speech_to_text.v1.websocket.BaseRecognizeCallback;
-
-/**
- * Recognize microphone input speech continuously using WebSockets.
- */
-public class MicrophoneWithWebSocketsExample {
-
-  /**
-   * The main method.
-   *
-   * @param args the arguments
-   * @throws Exception the exception
-   */
-  public static void main(final String[] args) throws Exception {
-    SpeechToText service = new SpeechToText();
-    service.setUsernameAndPassword("<username>", "<password>");
-
-    // Signed PCM AudioFormat with 16kHz, 16 bit sample size, mono
-    int sampleRate = 16000;
-    AudioFormat format = new AudioFormat(sampleRate, 16, 1, true, false);
-    DataLine.Info info = new DataLine.Info(TargetDataLine.class, format);
-
-    if (!AudioSystem.isLineSupported(info)) {
-      System.out.println("Line not supported");
-      System.exit(0);
-    }
-
-    TargetDataLine line = (TargetDataLine) AudioSystem.getLine(info);
-    line.open(format);
-    line.start();
-
-    AudioInputStream audio = new AudioInputStream(line);
-
-    RecognizeOptions options = new RecognizeOptions.Builder()
-        .audio(audio)
-        .interimResults(true)
-        .timestamps(true)
-        .wordConfidence(true)
-        //.inactivityTimeout(5) // use this to stop listening when the speaker pauses, i.e. for 5s
-        .contentType(HttpMediaType.AUDIO_RAW + ";rate=" + sampleRate)
-        .build();
-
-    service.recognizeUsingWebSocket(options, new BaseRecognizeCallback() {
-      @Override
-      public void onTranscription(SpeechRecognitionResults speechResults) {
-        System.out.println(speechResults);
-      }
-    });
-
-    System.out.println("Listening to your voice for the next 30s...");
-    Thread.sleep(30 * 1000);
-
-    // closing the WebSockets underlying InputStream will close the WebSocket itself.
-    line.stop();
-    line.close();
-
-    System.out.println("Fin.");
-  }
-}
