+++ examples/src/main/java/com/ibm/watson/text_to_speech/v1/TextToSpeechExample.java
@@ -1,0 +1,31 @@
+/**
+ * Copyright 2017 IBM Corp. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package com.ibm.watson.text_to_speech.v1;
+
+import com.ibm.cloud.sdk.core.service.security.IamOptions;
+import com.ibm.watson.developer_cloud.text_to_speech.v1.model.Voices;
+
+public class TextToSpeechExample {
+
+  public static void main(String[] args) {
+    TextToSpeech service = new TextToSpeech();
+    IamOptions options = new IamOptions.Builder()
+        .apiKey("<iam_api_key>")
+        .build();
+    service.setIamCredentials(options);
+
+    Voices voices = service.listVoices().execute().getResult();
+    System.out.println(voices);
+  }
+
+}
